split into SBMLImporter and SBMLParser
