+++ myokit/formats/sbml/__init__.py
@@ -7,12 +7,14 @@
 from __future__ import absolute_import, division
 from __future__ import print_function, unicode_literals
 
-from ._importer import SBMLImporter, SBMLError  # noqa
+from ._importer import SBMLImporter  # noqa
+from ._parser import SBMLError, SBMLParser
 
 
 # Importers
 _importers = {
     'sbml': SBMLImporter,
+    'sbml-parser': SBMLParser,
 }
